requirements.txt
asgiref==3.8.1
django==5.0.6
django-filter==24.2
django-taggit==5.0.1
djangorestframework==3.15.1
drf-nested-routers==0.94.1
pillow==10.3.0
sqlparse==0.5.0
typing-extensions==4.12.0
